Change framerate to float instead of int

+++ eitprocessing/eit_data/__init__.py
@@ -40,7 +40,7 @@ class EITData(ABC):
         path: PathLike | list[PathLike],
         vendor: Vendor | str,
         label: str | None = None,
-        framerate: int | None = None,
+        framerate: float | None = None,
         first_frame: int = 0,
         max_frames: int | None = None,
     ) -> Self:
